requirements.txt
streamlit
pandas
nltk
spacy
altair
openai-whisper
